Delete displayReservationSlice and refactor code
