+++ src/redux/reservationsSlice.js
@@ -1,14 +1,32 @@
-import { createSlice } from '@reduxjs/toolkit';
+import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
+import axios from 'axios';
 
-const reservationsSlice = createSlice({
-  name: 'reservations',
+export const getAllReservationsApi = createAsyncThunk(
+  'api/getAllReservationsApi',
+  async (userId) => {
+    const response = await axios.get(`http://localhost:3000/api/v1/users/${userId}/reservations`);
+    // console.log(response.data);
+    return response.data;
+  },
+);
+
+const displayreservationsSlice = createSlice({
+  name: 'api',
   initialState: {
-    reservations: [],
+    data: [],
+    status: null,
+    error: null,
+    userId: null,
   },
-  reducers: {
-  },
-  extraReducers: {
+
+  reducers: {},
+
+  extraReducers: (builder) => {
+    builder
+      .addCase(getAllReservationsApi.pending, (state) => ({ ...state, status: 'loading' }))
+      .addCase(getAllReservationsApi.fulfilled, (state, action) => ({ ...state, status: 'succeeded', data: action.payload }))
+      .addCase(getAllReservationsApi.rejected, (state, action) => ({ ...state, status: 'failed', error: action.error.message }));
   },
 });
 
-export default reservationsSlice.reducer;
+export default displayreservationsSlice.reducer;
